refactor: (crypto) split and re-organized `crypto.go` (#56)
* refactor: (crypto) extracted code from `crypto.go` into `beacon.go` and `rsa.go`

* fix the tests

---------
